Major refactor
Old brain is out! A new object or a stock, a stock group and a portfolio =
strategy to trade is created. Brain is cleaned and linked to initiate a
strategy portfolio and read the current state of the portfolio from file.

In the future some of the data, e.g, stocks id, boughtat and so on is to be
read from the brokers website. But portfolio risk and whitelist will still allways be stored on file.